Refactor cmsis-update into multiple modules
