+++ rust/cmsis-update/src/lib.rs
@@ -17,211 +17,47 @@ extern crate pdsc;
 
 use futures::prelude::*;
 use futures::Stream;
-use futures::stream::{futures_unordered, iter_ok};
-use hyper::{Body, Chunk, Client, Response, StatusCode, Uri};
+use hyper::{Body, Client};
 use hyper::client::Connect;
-use hyper::header::Location;
 use hyper_tls::HttpsConnector;
 use tokio_core::reactor::Core;
-use std::borrow::Borrow;
-use std::fs::OpenOptions;
-use std::io::Write;
 use std::iter::Iterator;
 use std::path::PathBuf;
 use clap::{App, ArgMatches, SubCommand};
 use slog::Logger;
 use failure::Error;
 
-use pack_index::{PdscRef, Pidx, Vidx};
 use pack_index::config::Config;
-use utils::parse::FromElem;
 
-trait ClientRedirExt<C>
-where
-    C: Connect,
-{
-    fn redirectable<'a>(
-        &'a self,
-        uri: Uri,
-        logger: &'a Logger,
-    ) -> Box<Future<Item = Response, Error = hyper::Error> + 'a>;
-}
+pub mod upgrade;
+mod redirect;
+mod vidx;
+mod dl_pdsc;
 
-impl<C: Connect> ClientRedirExt<C> for Client<C, Body> {
-    fn redirectable<'a>(
-        &'a self,
-        mut uri: Uri,
-        logger: &'a Logger,
-    ) -> Box<Future<Item = Response, Error = hyper::Error> + 'a> {
-        Box::new(async_block!{
-            let mut urls = Vec::new();
-            loop {
-                urls.push(uri.clone());
-                let res = await!(self.get(uri))?;
-                match res.status() {
-                    StatusCode::MovedPermanently |
-                    StatusCode::Found |
-                    StatusCode::SeeOther |
-                    StatusCode::TemporaryRedirect |
-                    StatusCode::PermanentRedirect => {
-                        let mut new_uri: Uri = res.headers()
-                            .get::<Location>()
-                            .unwrap_or(&Location::new(""))
-                            .parse()?;
-                        if let Some(ref old_uri) = urls.last() {
-                            if new_uri.authority().is_none() {
-                                if let Some(authority) = old_uri.authority() {
-                                    new_uri = format!("{}{}", authority, old_uri).parse()?
-                                }
-                            }
-                            debug!(logger, "Redirecting from {} to {}", old_uri, new_uri);
-                        }
-                        uri = new_uri;
-                    }
-                    _ => {
-                        return Ok(res);
-                    }
-                }
-            }
-        })
-    }
-}
+use vidx::{download_vidx_list, flatmap_pdscs};
+use dl_pdsc::{download_pdsc_stream};
 
-fn download_vidx<'a, C: Connect, I: Into<String>>(
-    client: &'a Client<C, Body>,
-    vidx_ref: I,
-    logger: &'a Logger,
-) -> impl Future<Item = Result<Vidx, minidom::Error>, Error = hyper::Error> + 'a {
-    let vidx = vidx_ref.into();
-    async_block!{
-        let uri = vidx.parse()?;
-        let body = await!(
-            client.redirectable(uri, logger)
-                .map(Response::body)
-                .flatten_stream()
-                .concat2())?;
-        Ok(parse_vidx(body, logger))
-    }
-}
 
-fn download_vidx_list<'a, C, I>(
-    list: I,
+/// Create a future of the update command.
+pub fn update_future<'a, C, I>(
+    config: &'a Config,
+    vidx_list: I,
     client: &'a Client<C, Body>,
     logger: &'a Logger,
-) -> impl Stream<Item = Result<Vidx, minidom::Error>, Error = hyper::Error> + 'a
-where
+) -> impl Future<Item = Vec<PathBuf>, Error = Error> + 'a
+    where
     C: Connect,
-    I: IntoIterator + 'a,
-    <I as IntoIterator>::Item: Into<String>,
+    I: IntoIterator<Item = String> + 'a,
 {
-    futures_unordered(
-        list.into_iter()
-            .map(|vidx_ref| download_vidx(client, vidx_ref, logger)),
-    )
-}
-
-fn parse_vidx(body: Chunk, logger: &Logger) -> Result<Vidx, minidom::Error> {
-    let string = String::from_utf8_lossy(body.as_ref());
-    Vidx::from_string(string.borrow(), logger)
-}
-
-fn into_uri(Pidx { url, vendor, .. }: Pidx) -> String {
-    format!("{}{}.pidx", url, vendor)
-}
-
-fn flatmap_pdscs<'a, C>(
-    Vidx {
-        vendor_index,
-        pdsc_index,
-        ..
-    }: Vidx,
-    client: &'a Client<C, Body>,
-    logger: &'a Logger,
-) -> impl Stream<Item = PdscRef, Error = Error> + 'a
-where
-    C: Connect,
-{
-    let pidx_urls = vendor_index.into_iter().map(into_uri);
-    let job = download_vidx_list(pidx_urls, client, logger)
-        .filter_map(|vidx| match vidx {
-            Ok(v) => Some(iter_ok(v.pdsc_index.into_iter())),
+    let parsed_vidx = download_vidx_list(vidx_list, client, logger);
+    let pdsc_list = parsed_vidx
+        .filter_map(move |vidx| match vidx {
+            Ok(v) => Some(flatmap_pdscs(v, client, logger)),
             Err(_) => None,
         })
         .flatten();
-    iter_ok(pdsc_index.into_iter()).chain(job)
-}
-
-fn make_uri(
-    &PdscRef {
-        ref url,
-        ref vendor,
-        ref name,
-        ..
-    }: &PdscRef,
-) -> Result<Uri, Error> {
-    let uri = if url.ends_with('/') {
-        format!("{}{}.{}.pdsc", url, vendor, name)
-    } else {
-        format!("{}/{}.{}.pdsc", url, vendor, name)
-    }.parse()?;
-    Ok(uri)
-}
-
-fn make_fd(
-    config: &Config,
-    &PdscRef {
-        ref vendor,
-        ref name,
-        ref version,
-        ..
-    }: &PdscRef,
-) -> PathBuf {
-    let mut filename = config.pack_store.clone();
-    let pdscname = format!("{}.{}.{}.pdsc", vendor, name, version);
-    filename.push(pdscname);
-    filename
-}
-
-fn download_pdsc<'a, C: Connect>(
-    config: &'a Config,
-    pdsc_ref: PdscRef,
-    client: &'a Client<C, Body>,
-    logger: &'a Logger,
-) -> impl Future<Item = Option<PathBuf>, Error = Error> + 'a {
-    async_block!{
-        let filename = make_fd(config, &pdsc_ref);
-        if filename.exists() {
-            return Ok(None);
-        }
-        let uri = make_uri(&pdsc_ref)?;
-        let PdscRef{vendor, name, version, ..} = pdsc_ref;
-        debug!(logger, "Updating package {}::{} to version {}", vendor, name, version);
-        let mut fd = OpenOptions::new()
-            .write(true)
-            .create(true)
-            .open(&filename)?;
-        let response = await!(client.redirectable(uri, logger))?;
-        #[async]
-        for bytes in response.body() {
-            fd.write_all(bytes.as_ref())?;
-        }
-        Ok(Some(filename))
-    }
-}
-
-fn download_pdsc_stream<'a, F, C>(
-    config: &'a Config,
-    stream: F,
-    client: &'a Client<C, Body>,
-    logger: &'a Logger,
-) -> impl Stream<Item = Option<PathBuf>, Error = Error> + 'a
-where
-    F: Stream<Item = PdscRef, Error = Error> + 'a,
-    C: Connect,
-{
-    stream
-        .map(move |pdsc_ref| download_pdsc(config, pdsc_ref, client, logger))
-        .buffer_unordered(32)
+    let pdscs = download_pdsc_stream(config, pdsc_list, client, logger);
+    pdscs.filter_map(id).collect()
 }
 
 fn id<T>(slf: T) -> T {
@@ -241,15 +77,7 @@ where
     C: Connect,
     I: IntoIterator<Item = String>,
 {
-    let parsed_vidx = download_vidx_list(vidx_list, client, logger);
-    let pdsc_list = parsed_vidx
-        .filter_map(|vidx| match vidx {
-            Ok(v) => Some(flatmap_pdscs(v, client, logger)),
-            Err(_) => None,
-        })
-        .flatten();
-    let pdscs = download_pdsc_stream(config, pdsc_list, client, logger);
-    core.run(pdscs.filter_map(id).collect())
+    core.run(update_future(config, vidx_list, client, logger))
 }
 
 /// Flatten a list of Vidx Urls into a list of updated CMSIS packs
